[MOL-12849] Move static map
