+++ src/components/shared/index.ts
@@ -3,3 +3,4 @@ export * from "./drag-upload";
 export * from "./error-messages";
 export * from "./prompt";
 export * from "./sanitize";
+export * from "./static-map";
