+++ src/components/shared/static-map/index.ts
