+++ src/components/shared/static-map/static-map.styles.ts
